Make time stoppable
Closes #21
